Move Api errors into their own file
